requirements.txt
requests==2.32.0
python-dotenv==1.1.1
pandas==2.1.0
openai
openpyxl==3.1.3
azure-ai-projects 
azure-identity